src/web-app/requirements.txt
Flask==2.2.3
azure-storage-blob==12.16.0
azure-identity==1.6.0